+++ resize/cimageresizer.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stdint.h>
 #include <type_traits>
 
 namespace ImageProcessing
@@ -9,8 +10,8 @@ namespace ImageProcessing
 	{
 		using DataPtr = std::conditional_t<ConstView, const void*, void*>;
 
-		uint32_t width;
-		uint32_t height;
+		uint64_t width;
+		uint64_t height;
 		uint8_t channels;
 		uint8_t bytesPerChannel;
 		uint8_t channelStride;
@@ -19,7 +20,7 @@ namespace ImageProcessing
 		DataPtr data;
 
 		template <class T>
-		[[nodiscard]] inline auto* scanLine(uint32_t line) const noexcept
+		[[nodiscard]] inline auto* scanLine(uint64_t line) const noexcept
 		{
 			if constexpr (ConstView)
 				return reinterpret_cast<const T*>(reinterpret_cast<const uint8_t*>(data) + static_cast<size_t>(line) * bytesPerLine);
@@ -27,10 +28,9 @@ namespace ImageProcessing
 				return reinterpret_cast<T*>(reinterpret_cast<uint8_t*>(data) + static_cast<size_t>(line) * bytesPerLine);
 		}
 	};
-
 	struct Rect {
-		uint32_t top = 0, left = 0;
-		uint32_t w = 0, h = 0;
+		uint64_t top = 0, left = 0;
+		uint64_t w = 0, h = 0;
 	};
 
 	void resize(ImageView<false>& dest, const ImageView<true>& source, Rect srcRect = {});
